Average Yield 2007-2012
During our meeting today we talked about obtaining an average crop yield
from year 2007 to 2012. Please take a look at this Excel file. We are
still missing quite a number of data--it appears that there are some
crops where data is unavailable for that particular crop type in a
particular location across 2007-2012. Please advise on what shall we do
to fill the missing data. Thanks.